Added unit tests using Jest

+++ utils/deleteTodo.js
@@ -1,12 +1,19 @@
-import { prisma } from "@/lib/prismaClient";
+import { prisma as prismaClient } from "@/lib/prismaClient";
 
-export default async function deleteTodo(todoId) {
-  const result = await prisma.todo.delete({
-    where: {
-      id: todoId,
-    },
-  });
-  return {
-    id: result?.id,
-  };
+export async function deleteTodo(todoId, prisma = prismaClient) {
+  if (todoId === null) {
+    return null;
+  }
+  try {
+    const result = await prisma.todo.delete({
+      where: {
+        id: todoId,
+      },
+    });
+    return {
+      id: result?.id,
+    };
+  } catch (err) {
+    return null;
+  }
 }
